+++ lib/compiler/fieldtype.dart
@@ -1,0 +1,10 @@
+/// This class describes Enum type declared in manifest.
+
+part of packme.compiler;
+
+class FieldType {
+    FieldType(this.filename, this.name);
+
+    final String filename;
+    final String name;
+}