+++ lib/compiler/parser.dart
@@ -3,101 +3,112 @@ part of packme.compiler;
 final Map<String, Enum> enums = <String, Enum>{};
 final Map<String, Message> types = <String, Message>{};
 final Map<int, Message> messages = <int, Message>{};
-final Map<int, Message> allMessages = <int, Message>{};
 
-bool _isEnumDeclaration(dynamic value) => value is List && value.isNotEmpty && value[0] is String;
-bool _isTypeDeclaration(dynamic value) => value is Map;
-bool _isMessageDeclaration(dynamic value) => value is List && value.length == 1 && value[0] is Map;
-bool _isRequestResponseDeclaration(dynamic value) => value is List && value.length == 2 && value[0] is Map && value[1] is Map;
-
-bool _nameDuplicated(String name) {
+bool _nameDuplicated(String name, [String? filename]) {
     for (final String key in enums.keys) if (name == key) return true;
     for (final String key in types.keys) if (name == key) return true;
-    for (final Message message in messages.values) if (name == message.name) return true;
+    for (final Message message in messages.values) {
+        if (name == message.name && (filename == null || message.filename == filename)) return true;
+    }
     return false;
 }
 
-void _checkIfValid(Map<String, dynamic> node) {
-    for (final MapEntry<String, dynamic> entry in node.entries) {
-        final String name = validName(entry.key, firstCapital: true);
-        if (name.isEmpty) throw Exception('Declaration name must be adequate :) "${entry.key}" is not.');
-        if (!_isEnumDeclaration(entry.value)
-            && !_isTypeDeclaration(entry.value)
-            && !_isMessageDeclaration(entry.value)
-            && !_isRequestResponseDeclaration(entry.value)) {
-            throw Exception('"${entry.key}" declaration is invalid. Use array of strings for enum declaration, object for type declaration or array (with 1 or 2 objects) for messages.');
+void _parseEnum(Node node) {
+    final String name = validName(node.name, firstCapital: true);
+    if (_nameDuplicated(name)) throw Exception('Enum "$name" in "${node.filename}" duplicates the name of another enum, type or message.');
+    for (final dynamic element in node.manifest) {
+        if (element is! String) throw Exception('Enum "$name" declaration must contain string values only.');
+    }
+    enums[node.name] = Enum(node.filename, name, (node.manifest as List<dynamic>).cast<String>());
+}
+
+void _parseType(Node node) {
+    final String name = validName(node.name, firstCapital: true);
+    if (_nameDuplicated(name)) throw Exception('Type "$name" in "${node.filename}" duplicates the name of another type, enum or message.');
+    types[node.name] = Message(node.filename, name, node.manifest as Map<String, dynamic>);
+}
+
+void _parseCommand(Node node) {
+    final String name = validName(node.name, firstCapital: true);
+    final String nameMessage = '${name}Message';
+    final String nameRequest = '${name}Request';
+    final String nameResponse = '${name}Response';
+    final int hashMessage = '${node.filename}$nameMessage'.hashCode;
+    final int hashRequest = '${node.filename}$nameRequest'.hashCode;
+    final int hashResponse = '${node.filename}$nameResponse'.hashCode;
+
+    if (node.type == NodeType.message) {
+        if (_nameDuplicated(nameMessage, node.filename)) {
+            throw Exception('Message "$nameMessage" in "${node.filename}" duplicates the name of another message, type or enum.');
         }
+        if (messages[hashMessage] != null) {
+            throw Exception('Message name "$nameMessage" in "${node.filename}" hash code turned out to be the same as for "${messages[hashMessage]!.name}". Please try another name.');
+        }
+        messages[hashMessage] = Message(node.filename, nameMessage, node.manifest[0] as Map<String, dynamic>, id: hashMessage);
+    }
+
+    if (node.type == NodeType.request) {
+        if (_nameDuplicated(nameRequest, node.filename)) {
+            throw Exception('Message "$nameRequest" in "${node.filename}" duplicates the name of another message, type or enum.');
+        }
+        if (_nameDuplicated(nameResponse, node.filename)) {
+            throw Exception('Message "$nameResponse" in "${node.filename}" duplicates the name of another message, type or enum.');
+        }
+        if (messages[hashRequest] != null) {
+            throw Exception('Message "$nameRequest" in "${node.filename}" name hash code turned out to be the same as for "${messages[hashRequest]!.name}". Please try another name.');
+        }
+        if (messages[hashResponse] != null) {
+            throw Exception('Message "$nameResponse" in "${node.filename}" name hash code turned out to be the same as for "${messages[hashResponse]!.name}". Please try another name.');
+        }
+        messages[hashResponse] = Message(node.filename, nameResponse, node.manifest[1] as Map<String, dynamic>, id: hashResponse);
+        messages[hashRequest] = Message(node.filename, nameRequest, node.manifest[0] as Map<String, dynamic>, id: hashRequest, responseClass: messages[hashResponse]);
     }
 }
 
-void _parseEnums(Map<String, dynamic> node, String prefix) {
-    for (final MapEntry<String, dynamic> entry in node.entries) {
-        final String name = validName(entry.key, firstCapital: true);
-        if (_isEnumDeclaration(entry.value)) {
-            if (_nameDuplicated(name)) throw Exception('Enum "$name" duplicates the name of another enum, type or message.');
-            for (final dynamic element in entry.value) {
-                if (element is! String) throw Exception('Enum "$name" declaration must contain string values only.');
+Map<String, List<String>> parse(List<Node> nodes) {
+    final Map<String, List<String>> codePerFile = <String, List<String>>{};
+    for (final Node node in nodes) codePerFile[node.filename] ??= <String>[];
+    nodes.where((Node node) => node.type == NodeType.enumeration).forEach(_parseEnum);
+    nodes.where((Node node) => node.type == NodeType.type).forEach(_parseType);
+    nodes.where((Node node) => node.type == NodeType.message || node.type == NodeType.request).forEach(_parseCommand);
+    for (final String filename in codePerFile.keys) {
+        codePerFile[filename] = <String>[
+            "import 'package:packme/packme.dart';",
+            '',
+            ...enums.values.where((Enum item) => item.filename == filename)
+                .fold(<String>[], (Iterable<String> a, Enum b) => a.toList() + b.output()),
+            ...types.values.where((Message item) => item.filename == filename)
+                .fold(<String>[], (Iterable<String> a, Message b) => a.toList() + b.output()),
+            ...messages.values.where((Message item) => item.filename == filename)
+                .fold(<String>[], (Iterable<String> a, Message b) => a.toList() + b.output()),
+            'final Map<int, PackMeMessage Function()> ${validName(filename)}MessageFactory = <int, PackMeMessage Function()>{',
+            ...messages.entries.where((MapEntry<int, Message> entry) => entry.value.filename == filename)
+                .map((MapEntry<int, Message> entry) => '${entry.key}: () => ${entry.value.name}.\$empty(),'),
+            '};'
+        ];
+
+        /// Add additional import lines according to messages' references data.
+        final List<FieldType> references = <FieldType>[];
+        for (final Message message in messages.values.where((Message item) => item.filename == filename)) {
+            references.addAll(message.references.where((FieldType reference) => !references.contains(reference)));
+        }
+        final Map<String, List<String>> imports = <String, List<String>>{};
+        for (final FieldType reference in references) {
+            /// Skip if enum/type declaration is in the same file.
+            if (reference.filename == filename) continue;
+            imports[reference.filename] ??= <String>[];
+            if (!imports[reference.filename]!.contains(reference.name)) {
+                imports[reference.filename]!.add(reference.name);
             }
-            enums[entry.key] = Enum(name, (entry.value as List<dynamic>).cast<String>());
+        }
+        final List<String> keys = imports.keys.toList();
+
+        /// Sort imports properly.
+        keys.sort((String a, String b) => '$a.generated.dart'.compareTo('$b.generated.dart'));
+        for (final String key in keys.reversed) {
+            imports[key]!.sort();
+            codePerFile[filename]!.insert(1, "import '$key.generated.dart' show ${imports[key]!.join(', ')};");
         }
     }
-}
-
-void _parseTypes(Map<String, dynamic> node, String prefix) {
-    for (final MapEntry<String, dynamic> entry in node.entries) {
-        final String name = validName(entry.key, firstCapital: true);
-        if (_isTypeDeclaration(entry.value)) {
-            if (_nameDuplicated(name)) throw Exception('Type "$name" duplicates the name of another type, enum or message.');
-            types[entry.key] = Message(name, entry.value as Map<String, dynamic>);
-        }
-    }
-}
-
-void _parseCommands(Map<String, dynamic> node, String prefix) {
-    for (final MapEntry<String, dynamic> entry in node.entries) {
-        final String name = validName(entry.key, firstCapital: true);
-        final String nameMessage = '${name}Message';
-        final String nameRequest = '${name}Request';
-        final String nameResponse = '${name}Response';
-        final int hashMessage = '$prefix$nameMessage'.hashCode;
-        final int hashRequest = '$prefix$nameRequest'.hashCode;
-        final int hashResponse = '$prefix$nameResponse'.hashCode;
-
-        if (_isMessageDeclaration(entry.value)) {
-            if (_nameDuplicated(nameMessage)) {
-                throw Exception('Message "$nameMessage" duplicates the name of another message, type or enum.');
-            }
-            if (allMessages[hashMessage] != null) {
-                throw Exception('Message name "$nameMessage" hash code turned out to be the same as for "${allMessages[hashMessage]!.name}". Please try another name.');
-            }
-            allMessages[hashMessage] = messages[hashMessage] = Message(nameMessage, entry.value[0] as Map<String, dynamic>, id: hashMessage);
-        }
-
-        else if (_isRequestResponseDeclaration(entry.value)) {
-            if (_nameDuplicated(nameRequest)) {
-                throw Exception('Message "$nameRequest" duplicates the name of another message, type or enum.');
-            }
-            if (_nameDuplicated(nameResponse)) {
-                throw Exception('Message "$nameResponse" duplicates the name of another message, type or enum.');
-            }
-            if (allMessages[hashRequest] != null) {
-                throw Exception('Message name "$nameRequest" hash code turned out to be the same as for "${allMessages[hashRequest]!.name}". Please try another name.');
-            }
-            if (allMessages[hashResponse] != null) {
-                throw Exception('Message name "$nameResponse" hash code turned out to be the same as for "${allMessages[hashResponse]!.name}". Please try another name.');
-            }
-            allMessages[hashResponse] = messages[hashResponse] = Message(nameResponse, entry.value[1] as Map<String, dynamic>, id: hashResponse);
-            allMessages[hashRequest] = messages[hashRequest] = Message(nameRequest, entry.value[0] as Map<String, dynamic>, id: hashRequest, responseClass: messages[hashResponse]);
-        }
-    }
-}
-
-void parse(Map<String, dynamic> node, String prefix) {
-    enums.clear();
-    types.clear();
-    messages.clear();
-    _checkIfValid(node);
-    _parseEnums(node, prefix);
-    _parseTypes(node, prefix);
-    _parseCommands(node, prefix);
+    return codePerFile;
 }